JibconLoginManager Singleton화 / Logout정상화 (#92) JibconLoginManager 인터페이스 제거 JibconLoginManager 싱글톤 외에 인스턴트 제작하는부분 삭제 Logout후에 mUser = null로 변경 SharedPref에 저장하는부분에 쓰인 문자열들 Constant화
